requirements.txt
aniso8601==10.0.0
APScheduler==3.11.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
eventlet==0.39.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-Session==0.5.0
Flask-SocketIO==5.3.4
fonttools==4.57.0
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
holidays==0.70
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mysql-connector-python==8.0.33
nepali-datetime==1.0.8.4
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
prophet==1.1.6
protobuf==3.20.3
PyJWT==2.7.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.7.2
pytz==2023.3
requests==2.32.3
simple-websocket==0.10.1
six==1.17.0
stanio==0.5.1
tqdm==4.67.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websocket-client==1.5.3
Werkzeug==2.3.6
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
celery==5.3.1
redis==4.5.3
sendgrid==6.10.0
yagmail==0.15.293
